Vision-Parking/tests/requirements.txt
appium-python-client>=2.11.0
pytest>=7.0.0
pytest-html>=3.1.0
selenium>=4.0.0
Pillow>=9.0.0